feature: shopping cart, place order, send gmail for customer when order success; fix: login-logout admin;

+++ database/migrations/2021_08_09_050359_create_order_detail_table.php
@@ -18,9 +18,9 @@ class CreateOrderDetailTable extends Migration
             $table->unsignedBigInteger("order");
             $table->unsignedBigInteger("product");
             $table->integer("quantity",false,true)->default(0);
-            $table->integer("price",false,true)->default(0);
-            $table->integer("discounted",false,true)->default(0);
-            $table->integer("final_price",false,true)->default(0);
+            $table->double("price",false,true)->default(0);
+            $table->double("discounted",false,true)->default(0);
+            $table->double("final_price",false,true)->default(0);
             $table->timestamps();
             $table->softDeletes();
         });
